requirements.txt
streamlit>=1.32.0
plotly>=5.20.0
pandas>=2.2.0
yfinance>=0.2.37
requests>=2.31.0
openpyxl>=3.1.2
python-dotenv>=1.0.1
kaleido>=0.2.1
